desktop/requirements.txt
fastapi>=0.109.1
uvicorn>=0.24.0
pyautogui>=0.9.54
Pillow>=10.2.0
numpy>=1.24.0
firebase-admin>=6.2.0
pyngrok>=7.0.0
requests>=2.31.0
pygetwindow>=0.0.9
